Split mount option validation in its own package
