+++ pkg/registry/core/persistentvolume/strategy.go
@@ -29,6 +29,7 @@ import (
 	"k8s.io/apiserver/pkg/storage/names"
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/validation"
+	volume_validation "k8s.io/kubernetes/pkg/volume/validation"
 )
 
 // persistentvolumeStrategy implements behavior for PersistentVolume objects
@@ -53,7 +54,8 @@ func (persistentvolumeStrategy) PrepareForCreate(ctx genericapirequest.Context, 
 
 func (persistentvolumeStrategy) Validate(ctx genericapirequest.Context, obj runtime.Object) field.ErrorList {
 	persistentvolume := obj.(*api.PersistentVolume)
-	return validation.ValidatePersistentVolume(persistentvolume)
+	errorList := validation.ValidatePersistentVolume(persistentvolume)
+	return append(errorList, volume_validation.ValidateForMountOptions(persistentvolume)...)
 }
 
 // Canonicalize normalizes the object after validation.
@@ -72,8 +74,10 @@ func (persistentvolumeStrategy) PrepareForUpdate(ctx genericapirequest.Context, 
 }
 
 func (persistentvolumeStrategy) ValidateUpdate(ctx genericapirequest.Context, obj, old runtime.Object) field.ErrorList {
-	errorList := validation.ValidatePersistentVolume(obj.(*api.PersistentVolume))
-	return append(errorList, validation.ValidatePersistentVolumeUpdate(obj.(*api.PersistentVolume), old.(*api.PersistentVolume))...)
+	newPv := obj.(*api.PersistentVolume)
+	errorList := validation.ValidatePersistentVolume(newPv)
+	errorList = append(errorList, volume_validation.ValidateForMountOptions(newPv)...)
+	return append(errorList, validation.ValidatePersistentVolumeUpdate(newPv, old.(*api.PersistentVolume))...)
 }
 
 func (persistentvolumeStrategy) AllowUnconditionalUpdate() bool {
